+++ src/components/vetdoc/index.docvetcomp.tsx
@@ -1,11 +1,8 @@
-export * from './view/Analisis';
-export * from './view/Receta';
 export * from './view/ServiciosActivos';
 export * from './view/ServiciosSeleccion';
 export * from './view/Sidebar';
 export * from './view/VetHeader';
 
-// export * from './CompletarConsultaModal';
 export * from './formularios/CirugiaForm';
 export * from './formularios/ConsultaForm';
 export * from './formularios/InternacionForm';
@@ -13,6 +10,11 @@ export * from './formularios/PeluqueriaForm';
 export * from './formularios/VacunaForm';
 export * from './formularios/VacunationForm'
 ;
+export * from './listados/Analisis';
+export * from './listados/Receta';
+export * from './listados/VacunaList';
+export * from './listados/VacunationList';
+
 export * from './modals/CompletarCirugiaModal';
 export * from './modals/CompletarConsultaModal';
 export * from './modals/CompletarInternacionModal';
@@ -20,5 +22,3 @@ export * from './modals/CompletarInternacionModal';
 export * from './ReservacionesPendientes';
 export * from './ServiciosActivos';
 export * from './ServiciosCompletados';
-export * from './VacunaList';
-export * from './VacunationList';
