+++ src/components/vetdoc/listados/Receta.tsx
